not delete theme

+++ ajax/cache/global.php
@@ -3,11 +3,12 @@
 $dir = PATH_HOME . (DEV ? "assetsPublic" : "assets");
 foreach(new RecursiveIteratorIterator(new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS),
     RecursiveIteratorIterator::CHILD_FIRST) as $file) {
-    if ($file->isDir())
-        rmdir($file->getRealPath());
-    else
-        unlink($file->getRealPath());
+    if(!in_array($file->getFileName(), ["theme.css", "theme"])) {
+        if ($file->isDir())
+            rmdir($file->getRealPath());
+        elseif ($file->getFileName())
+            unlink($file->getRealPath());
+    }
 }
-rmdir($dir);
 
 $data['data'] = "1";
